+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-02T06:41:47.199Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-02T07:25:06.598Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")